pyproject.toml
[project]
name = "jobfinder"
version = "0.1.0"
description = "CLI tool to discover relevant companies and open roles from your resume"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "anthropic>=0.42",
    "google-genai>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.scripts]
jobfinder = "jobfinder.cli:cli"

[tool.setuptools.packages.find]
include = ["jobfinder*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
